backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-pinecone==0.0.1
langsmith>=0.0.77,<0.1.0
pinecone-client>=3,<4
motor==3.3.2
pymongo==4.6.1
pypdf==3.17.4
python-docx==1.1.0
python-dotenv==1.0.0
sentence-transformers>=2.3.0
openai==1.6.1
ollama==0.1.4
httpx==0.25.2
aiofiles==23.2.1

